+++ sdk/lib/_internal/vm/lib/regexp_patch.dart
@@ -117,7 +117,7 @@ class _RegExpMatch implements RegExpMatch {
 
   int get groupCount => _regexp._groupCount;
 
-  Pattern get pattern => _regexp;
+  RegExp get pattern => _regexp;
 
   String? namedGroup(String name) {
     var idx = _regexp._groupNameIndex(name);