+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 #### `dart:core`
 - Added `bool.parse` and `bool.tryParse` static methods.
 - Added `DateTime.timestamp()` constructor to get current time as UTC.
+- The type of `RegExpMatch.pattern` is now `RegExp`, not just `Pattern`.
 
 - **Breaking change** [#49529][]:
   - Removed the deprecated `List` constructor, as it wasn't null safe.